trying fixes again, for 23rd. i think the problem was making montlake fixes, then more huge_seattle fixes that invalidated the montlake fixes. it's not hard to do the last-minute border trimming, so be willing to restart that.
but actually, also have to skip deleting intersections if roads are
connected. urgh.

+++ map_model/src/raw.rs
@@ -135,7 +135,12 @@ impl RawMap {
 
         for i in &fixes.delete_intersections {
             if self.intersections.contains_key(i) {
-                self.delete_intersection(*i);
+                if self.can_delete_intersection(*i) {
+                    self.delete_intersection(*i);
+                } else {
+                    println!("Can't delete {}, roads are connected, skipping", i);
+                    skipped += 1;
+                }
                 applied += 1;
             } else {
                 skipped += 1;
